+++ src/client/index.ts
@@ -810,11 +810,24 @@ export class SunsamaClient {
       duration: null,
       runDate: null,
       snooze,
-      timeHorizon: null,
+      timeHorizon: options?.timeHorizon
+        ? {
+            type: options.timeHorizon,
+            relativeTo: new Date().toISOString(),
+          }
+        : null,
       dueDate,
       comments: [],
       orderings: [],
-      backlogOrderings: [],
+      backlogOrderings: options?.timeHorizon
+        ? [
+            {
+              horizonType: options.timeHorizon,
+              position: 0,
+              streamId: options?.streamIds?.[0] || null,
+            },
+          ]
+        : [],
       subtasks: [],
       subtasksCollapsed: null,
       sequence: null,