+++ src/server/utils/index.ts
@@ -1,0 +1,5 @@
+/**
+ * Server utilities exports
+ */
+
+export * from './logger.js';